02-02-2024, rutinas

+++ app/src/main/java/com/example/proyectapp/ExerciseAdapter.kt
@@ -1,20 +1,23 @@
 package com.example.proyectapp
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.RecyclerView
 import com.example.proyectapp.database.AppDatabase
 import com.example.proyectapp.databinding.ExerciseLayoutBinding
+import com.example.proyectapp.model.Exercise
 
 
 class ExerciseAdapter (
     var exercises: List<Exercise>,
     val context: Context,
     val db: AppDatabase
+
 ) :
 
     RecyclerView.Adapter<ExerciseAdapter.ItemViewHolder>() {
@@ -48,22 +51,49 @@ class ExerciseAdapter (
 
         binding.texttViewNotes.text = exercise.notes
         binding.buttonDelete.setOnClickListener{
-            val deletedRows = db.exerciseDao().delete(exercise.id)
+            mostrarDialogoConfirmacion { confirmacion ->
+                var deletedRows = 0
+                if(confirmacion){
+                    deletedRows = db.exerciseDao().delete(exercise.id)
+                    exercises = db.exerciseDao().list()
+                    notifyDataSetChanged()
+                }
 
-            exercises = db.exerciseDao().list()
-
-            notifyDataSetChanged()
-            if(deletedRows == 0) {
-                Toast.makeText(context, "No se ha eliminado ningún ejercicio", Toast.LENGTH_LONG).show()
-            }else {
-                Toast.makeText(context, "Se ha eliminado el ejercicio", Toast.LENGTH_LONG).show()
+                if(deletedRows == 0) {
+                    Toast.makeText(context, "No se ha eliminado ningún ejercicio", Toast.LENGTH_LONG).show()
+                } else {
+                    Toast.makeText(context, "Se ha eliminado el ejercicio", Toast.LENGTH_LONG).show()
+                }
             }
         }
 
+        binding.buttonEdit.setOnClickListener{
 
 
-
+        }
     }
+
+    private fun mostrarDialogoConfirmacion(callback: (Boolean) -> Unit) {
+        val builder = AlertDialog.Builder(context)
+
+        builder.setTitle("Confirmar eliminación")
+        builder.setMessage("¿Estás seguro de que quieres realizar esta acción?")
+
+        builder.setPositiveButton("Sí") { dialogInterface: DialogInterface, _: Int ->
+            callback(true)
+            dialogInterface.dismiss()
+        }
+
+        builder.setNegativeButton("No") { dialogInterface: DialogInterface, _: Int ->
+            callback(false)
+            dialogInterface.dismiss()
+        }
+
+        val dialog = builder.create()
+        dialog.show()
+    }
+
+
 
 
 }
